Result class를 통한 refactoring

+++ src/main/java/kr/jason/web/QuestionController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import kr.jason.domain.Question;
 import kr.jason.domain.QuestionRepository;
+import kr.jason.domain.Result;
 import kr.jason.domain.User;
 
 
@@ -50,15 +51,25 @@ public class QuestionController {
 	}
 	@GetMapping("/{id}/form")
 	public String updateForm(@PathVariable Long id, Model model, HttpSession session){
-		try{
-			Question question = questionRepository.findOne(id);
-			hasPermission(session, question);
-			model.addAttribute("question", question);
-			return "/qna/updateForm";
-		}catch(IllegalStateException e){
-			model.addAttribute("errorMessage", e.getMessage());
-			return "/user/login";	
+		Question question = questionRepository.findOne(id);
+		Result result = valid(session, question);
+		if(!result.isValid()){
+			model.addAttribute("errorMessage", result.getErrorMessage());
+			return "/user/login";			
 		}
+		model.addAttribute("question", question);
+		return "/qna/updateForm";
+	}
+	
+	private Result valid(HttpSession session, Question question){
+		if(!HttpSessionUtils.isLoginUser(session)){
+			return Result.fail("로그인이 필요합니다.");
+		}
+		User logindUser = HttpSessionUtils.getUserFromSession(session);
+		if(!question.isSameUser(logindUser)){
+			return Result.fail("작성자만 수정/삭제 가능합니다.");
+		}
+		return Result.ok();
 	}
 	
 	private boolean hasPermission(HttpSession session, Question question){
@@ -74,27 +85,27 @@ public class QuestionController {
 	//@RequestMapping(value="/questions", method=RequestMethod.PUT)
 	@PutMapping("/{id}")
 	public String update(@PathVariable Long id, Model model, String title, String contents, HttpSession session){
-		try{
-			Question question = questionRepository.findOne(id);
-			hasPermission(session, question);
-			question.update(title,contents);
-			questionRepository.save(question);
-			return String.format("redirect:/questions/%d",id);
-		}catch(IllegalStateException e){
-			model.addAttribute("errorMessage", e.getMessage());
-			return "/user/login";	
+		Question question = questionRepository.findOne(id);
+		Result result = valid(session, question);
+		if(!result.isValid()){
+			model.addAttribute("errorMessage", result.getErrorMessage());
+			return "/user/login";			
 		}
+		question.update(title,contents);
+		questionRepository.save(question);
+		return String.format("redirect:/questions/%d",id);
+	
 	}
 	@DeleteMapping("/{id}")
 	public String delete(@PathVariable Long id, Model model, HttpSession session){
-		try{
-			Question question = questionRepository.findOne(id);
-			hasPermission(session, question);
-			questionRepository.delete(id);
-			return "redirect:/";
-		}catch(IllegalStateException e){
-			model.addAttribute("errorMessage", e.getMessage());
-			return "/user/login";	
+		Question question = questionRepository.findOne(id);
+		Result result = valid(session, question);
+		if(!result.isValid()){
+			model.addAttribute("errorMessage", result.getErrorMessage());
+			return "/user/login";			
 		}
+		questionRepository.delete(id);
+		return "redirect:/";
 	}
+	
 }
